Revert "Changed code reflected Chura Data comments"
