+++ app/channels/roomcontrol_channel.rb
@@ -1,6 +1,7 @@
 require 'base64'
 
 class RoomcontrolChannel < ApplicationCable::Channel
+  include Euchre
 
   def subscribed
     stream_from "chat_#{params[:room_id]}"
