+++ app/controllers/pages_controller.rb
@@ -2,7 +2,7 @@ class PagesController < ApplicationController
   include ApplicationHelper
   def index
     redirect_to '/game/new' if logged_in?
-    @updates = GameUpdates.last(5).reverse!
+    @updates = Gameupdates.last(5).reverse!
   end
 
 end
