--- lib/euchre/deck.rb
@@ -1,27 +1,0 @@
-#generate euchre deck here with shuffle function
-class Deck
-    attr_accessor :cards
-
-    def initialize
-      @cards = []
-      suits = (0..3).to_a
-      #only include ace, 9s,10s,Jacks,Queens, and kings
-      values = [0,8,9,10,11,12]
-      suits.each do |suit|
-        values.each do |value|
-          card = Card.new(suit,value)
-          @cards.push(card)
-        end
-      end
-      @cards.shuffle!
-    end
-
-    def shuffle
-      @cards.shuffle!
-    end
-
-    def deal_card
-      return @cards.pop
-    end
-
-  end
